backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
